jsonschema+jmespath/pyproject.toml
[tool.poetry]
name = "jsonschema+jmespath"
version = "0.1.0"
description = ""
authors = ["Michal Opala <sk4zuzu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jsonschema = "^3.2.0"
jmespath = "^0.10.0"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
